+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add `read.ctd.saiv()` to read data from SAIV instruments (issue 2141).
 * Remove many dontrun examples from the docs (issue 2152).
 * Remove `trimString()`, after it had been deprecated for a long time (issue 2123).
+* Remove use of the deprecated `sp` package (issue 2154).
 * Change `amsr` dataset and functions to handle new format (issues 2124 to 2133, 2147).
 * Change `as.tidem()` to handle NOAA constituent names and frequencies (issues 2143, 2144 and 2146).
 * Change `computableWaterProperties` to handle a vector of variable names.